Add synth project with sine osc, filter, and more
Add my first synthesizer project (based on JUCE's "audio plugin demo" to
the repo.

This already has all of the audio processing that's done in the JUCE
demo, plus my Transposed Direct Form II Filter. Also, the newer
AudioProcessorParameter class is in use and all of my parameters are
created using it. So, the PluginProcessor files are pretty much done
(except for custom changes that will be made later).

I still need to create a GUI with things like Sliders and a ComboBox so
I can "connect" to the AudioProcessor and control the
AudioProcessorParameters from the UI.

Note:
This follows the recently updated "plugin host demo", which uses a whole
new way of dealing with parameters. Make sure to remember not to use the
old way of handling them!