Add GUI, custom paramer classes, etc.
Add GUI components like Sliders and a ComboBox for the needed/expected
controls. Also, add a MidiKeyboardComponent that has a graphical
keyboard that acts as a virtual keyboard of some sort.

Create custom AudioProcessorParameter classes for each parameter of the
filter (other unused parameters still needed). Also, set up the things
like the timer callback that updates the UI and others to change the
value of the parameters from the UI.

Update filter to have more reasonable member functions/arguments. Also,
create one function that can process in stereo (2 channels), instead of
using two separate functions. This is done using arrays for the states
(z^-1).

Note:
The filterType ComboBox isn't working properly and needs to be fixed so
it'll work with the AudioProcessorParameter system employed by other
parameters.