Create saw/ramp oscillator
Create a sawtooth/ramp oscillator that can be controlled with MIDI and
can be used polyphonically, since it uses JUCE's SynthesiserVoice and
SynthesiserSound classes. This could also be used as a phase accumulator
that could be used with other oscillators, with very slight
modification.

Note:
This still needs a way to change the pitch (semitones/cents). This also
needs a way of changing the phase shift. These could both be done
through member functions like most other classes.